Changes in content in products array, create with class
I changed the products array to create it from a class, so that it's more dynamic to add content and change it.
I changed the descriptions so that the text are not too short and can be tested in the design of the product item presentation.
I add for future use "id" and "quantity" so that to use it in the cart and checkout pages.